Rename function and files for better clarity
The function 'generateFullMarkdown' was renamed to 'convertJsonSchemasToMarkdowns' for better accuracy of its functionality. Additionally, several files ('sql-schema.js', 'array-shape.js') were renamed ('schema-sql.js', 'schema-array-shape.js') to improve clarity and project organization.
