+++ index.html
@@ -97,8 +97,8 @@
   <div id="option2Content" class="preview-content" style="display: none;"></div>
 </div>
 
-<script src="array-shape.js"></script>
-<script src="sql-schema.js"></script>
+<script src="schema-array-shape.js"></script>
+<script src="schema-sql.js"></script>
 <script src="schema-markdown.js"></script>
 <script>
   function autoResize(textarea) {
