+++ schema-array-shape.js
